update social login
