--- app/Http/Controllers/Auth/API/LoginController.php
@@ -1,181 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\Auth\API;
-
-use App\Http\Controllers\Controller;
-use App\Providers\SocialAccountService;
-use Illuminate\Foundation\Auth\AuthenticatesUsers;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Lang;
-use Laravel\Socialite\Facades\Socialite;
-use Tymon\JWTAuth\Facades\JWTAuth;
-
-class LoginController extends Controller
-{
-    /*
-    |--------------------------------------------------------------------------
-    | Login Controller
-    |--------------------------------------------------------------------------
-    |
-    | This controller handles authenticating users for the application and
-    | redirecting them to your home screen. The controller uses a trait
-    | to conveniently provide its functionality to your applications.
-    |
-    */
-
-    use AuthenticatesUsers;
-
-    /**
-     *  List of supported social providers.
-     */
-    private $socialProviders = ['google', 'facebook'];
-
-    /**
-     * Create a new controller instance.
-     */
-    public function __construct()
-    {
-        $this->middleware('guest', ['except' => 'logout']);
-    }
-
-    /**
-     *  Redirect to social provider
-     *
-     * @param $provider
-     * @return \Illuminate\Http\JsonResponse
-     */
-    public function redirectToProvider($provider)
-    {
-        if(! in_array($provider, $this->socialProviders)){
-            return response()->json([
-                'message' => Lang::get('auth.social.failed'),
-            ], 401);
-        }
-
-        return Socialite::driver($provider)->redirect();
-    }
-
-    /**
-     * Handle retrieved information from provider
-     *
-     * @param $provider
-     * @param SocialAccountService $accountService
-     * @param Request $request
-     * @return \Illuminate\Http\JsonResponse|\Illuminate\Http\RedirectResponse
-     */
-    public function handleProviderCallback($provider, SocialAccountService $accountService, Request $request)
-    {
-        $socialUser = Socialite::driver($provider)->user();
-
-        $user = $accountService->getUser($socialUser, $provider);
-
-        if ($user) {
-            $token = JWTAuth::fromUser($user);
-
-            return $this->sendLoginResponse($request, $token);
-        }
-
-        return $this->sendFailedLoginResponse($request);
-    }
-
-    /**
-     * Handle a login request to the application.
-     *
-     * @param Request $request
-     * @return \Illuminate\Http\JsonResponse|\Illuminate\Http\RedirectResponse
-     */
-    public function login(Request $request)
-    {
-        $this->validateLogin($request);
-
-        // If the class is using the ThrottlesLogins trait, we can automatically throttle
-        // the login attempts for this application. We'll key this by the username and
-        // the IP address of the client making these requests into this application.
-        if ($this->hasTooManyLoginAttempts($request)) {
-            $this->fireLockoutEvent($request);
-
-            return $this->sendLockoutResponse($request);
-        }
-
-        $credentials = $this->credentials($request);
-
-        if ($token = JWTAuth::attempt($credentials)) {
-            return $this->sendLoginResponse($request, $token);
-        }
-
-        // If the login attempt was unsuccessful we will increment the number of attempts
-        // to login and redirect the user back to the login form. Of course, when this
-        // user surpasses their maximum number of attempts they will get locked out.
-        $this->incrementLoginAttempts($request);
-
-        return $this->sendFailedLoginResponse($request);
-    }
-
-    /**
-     * Send the response after the user was authenticated.
-     *
-     * @param Request $request
-     * @param string $token
-     * @return \Illuminate\Http\JsonResponse
-     */
-    protected function sendLoginResponse(Request $request, string $token)
-    {
-        $this->clearLoginAttempts($request);
-
-        return $this->authenticated($token);
-    }
-
-    /**
-     * Get the failed login response instance.
-     *
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    protected function sendFailedLoginResponse(Request $request)
-    {
-        return response()->json([
-            'message' => Lang::get('auth.failed'),
-        ], 401);
-    }
-
-    /**
-     * Send the response after the user was authenticated.
-     *
-     * @param Request $request
-     * @param $user
-     * @param string $token
-     * @return \Illuminate\Http\JsonResponse
-     */
-    protected function authenticated(string $token)
-    {
-        return response()->json([
-            'token' => $token,
-        ]);
-    }
-
-    /**
-     * Get the login username to be used by the controller.
-     *
-     * @return string
-     */
-    public function username()
-    {
-        return 'email';
-    }
-
-    /**
-     * Log the user out of the application.
-     *
-     * @param \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function logout(Request $request)
-    {
-        $this->guard()->logout();
-
-        $request->session()->flush();
-
-        $request->session()->regenerate();
-
-        return response()->json(['message' => Lang::get('auth.logged.out')]);
-    }
-}
