+++ app/Http/Controllers/Users/UsersController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\API\Users;
+namespace App\Http\Controllers\Users;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
@@ -49,6 +49,6 @@ class UsersController extends Controller
      */
     public function destroy($id)
     {
-        //
+
     }
 }
